Add files via upload
In Week 1, the primary objective was to clean messy data set. The task was to understand the dataset, clean it professionally, and prepare it for analysis. 
•	I created a new sheet within the Excel file and renamed it Cleaned sheet. This new sheet served as the working area for all data transformation and preparation tasks. I removed unwanted columns and empty cells, ensuring that only relevant and complete information remained. This step helped to streamline the dataset, making it more manageable and focused for subsequent analysis. 
•	Next, I rearranged the remaining columns to present the data in a more logical and coherent order, enhancing the overall narrative and interpretability of the dataset. To enrich the cleaned data further, I utilised lookup functions such as VLOOKUP and XLOOKUP. These functions were particularly valuable in populating important information into columns that were initially blank. By referencing other sheets provided within the same Excel file, I was able to accurately fill these gaps, ensuring the dataset was as complete and informative as possible. 
Through a systematic approach to cleaning, rearranging, and enhancing the dataset, I successfully prepared the data for subsequent analysis, laying a solid foundation for the next phase of the project.


In Week 2, the emphasis was the creation and refinement of calculated columns within the cleaned dataset. Several new columns were introduced with the aim of providing clearer and more actionable insights for analysis.
Among the columns added were:
•	Total Sales
•	Total Sales Before Discount
•	Total Sales After Discount
•	Discount Amount
•	Discount Percentage
•	Number of Days to Ship
•	Shipping Speed Category
•	Order Category
•	Discount Level
Each column was carefully designed to capture specific aspects of the data, enabling more detailed and meaningful interpretation. For instance, calculating discount amounts and percentages provided a clear view of promotional impacts, while categorising shipping speeds and orders allowed for segmented analysis.
In addition to constructing these columns, attention was given to professional formatting. The appropriate currency format was applied to all price-related columns, discount figures were formatted accordingly, and numerical values were standardised throughout. This ensured consistency and readability, making the dataset easier to work with and interpret during subsequent analysis.